Fixed some compiler warnings and errors in test suite found by buildbot
include/ft_global.h:
  Changed string size to 'size_t' (to avoid compiler warnings later)
include/maria.h:
  Changed string size to 'size_t' (to avoid compiler warnings later)
mysql-test/r/show_check.result:
  Don't use logging to tables, as there is no guarantee in which order tables are showed in 'show open tables'
mysql-test/suite/rpl/t/rpl_rotate_logs.test:
  Use HIGH_PRIORITY inserts to avoid problem with concurrent inserts on the slave.
mysql-test/t/show_check-master.opt:
  Don't use logging to tables, as there is no guarantee in which order tables are showed in 'show open tables'
sql/item.cc:
  Fixed compiler warning
storage/maria/ft_maria.c:
  Changed string size to 'size_t' (to avoid compiler warnings later)
storage/myisam/ft_boolean_search.c:
  Fixed compiler warnings
storage/myisam/ft_myisam.c:
  Changed string size to 'size_t' (to avoid compiler warnings later)
storage/myisam/ft_nlq_search.c:
  Changed string size to 'size_t' (to avoid compiler warnings later)
storage/myisam/ft_parser.c:
  Changed string size to 'size_t' (to avoid compiler warnings)
storage/myisam/ft_stopwords.c:
  Changed string size to 'size_t' (to avoid compiler warnings)
storage/myisam/ftdefs.h:
  Changed string size to 'size_t' (to avoid compiler warnings)

+++ include/maria.h
@@ -462,7 +462,7 @@ void maria_versioning(MARIA_HA *info, my_bool versioning);
 void maria_ignore_trids(MARIA_HA *info);
 
 /* fulltext functions */
-FT_INFO *maria_ft_init_search(uint,void *, uint, uchar *, uint,
+FT_INFO *maria_ft_init_search(uint,void *, uint, uchar *, size_t,
                               CHARSET_INFO *, uchar *);
 
 /* 'Almost-internal' Maria functions */
